getTaxonInfo gets trait information for a taxon getTaxonomy gets taxonomic information for a taxon metadata gets the description of the content of a database.

getTaxonInfo extracts trait information for a certain taxon from a trait database.

metadata returns a description of an object, i.e the meaning of the columns, units, etc...

Usage

getTaxonInfo(taxon, trait = Traits_nioz, taxonomy = Taxonomy)  

getTaxonomy(taxon, taxonomy=Taxonomy)

metadata(object)

Arguments

taxon

name of the taxon (or vector with names) for which the traits need to be listed.

trait

a trait database in wide format; it should have taxon as first column, and it should contain a description of the trait database in its attribute ("description"). See Traits_nioz for a suitable structure.

taxonomy

taxonomic information; first column will be matched with taxon from the trait database, regardless of its name.

object

any object that has a data.frame description in its attributes.

Value

getTaxonInfo returns a data.frame with the traits of the species whose modalities are not Null, NULL or None.

getTaxonomy returns the occurrences of taxon within taxonomy.

metadata returns a data.frame with a description of the contents of the database.

Author

Karline Soetaert <karline.soetaert@nioz.nl> Olivier Beauchard

See also

Examples

## ====================================================
## Description of the Traits_nioz database
## ====================================================
metadata(Traits_nioz)
#>     colname                            trait                modality indic
#> 1    ET1.M1    Substratum depth distribution                       0     1
#> 2    ET1.M2    Substratum depth distribution                     0-5     1
#> 3    ET1.M3    Substratum depth distribution                    5-15     1
#> 4    ET1.M4    Substratum depth distribution                   15-30     1
#> 5    ET1.M5    Substratum depth distribution                     >30     1
#> 6    ET2.M1                     Biodiffusion                    Null     2
#> 7    ET2.M2                     Biodiffusion                     Low     2
#> 8    ET2.M3                     Biodiffusion                    High     2
#> 9    ET3.M1               Downward conveying                    Null     3
#> 10   ET3.M2               Downward conveying                     Low     3
#> 11   ET3.M3               Downward conveying                    High     3
#> 12   ET4.M1                 Upward conveying                    Null     4
#> 13   ET4.M2                 Upward conveying                     Low     4
#> 14   ET4.M3                 Upward conveying                    High     4
#> 15   ET5.M1                     Regeneration                    Null     5
#> 16   ET5.M2                     Regeneration                     Low     5
#> 17   ET5.M3                     Regeneration                    High     5
#> 18   ET6.M1                    Biodeposition                    Null     6
#> 19   ET6.M2                    Biodeposition                     Low     6
#> 20   ET6.M3                    Biodeposition                    High     6
#> 21   ET7.M1                       Bioerosion                    Null     7
#> 22   ET7.M2                       Bioerosion                     Low     7
#> 23   ET7.M3                       Bioerosion                    High     7
#> 24   ET8.M1                 Biostabilisation                    Null     8
#> 25   ET8.M2                 Biostabilisation                     Low     8
#> 26   ET8.M3                 Biostabilisation                    High     8
#> 27   ET9.M1              Ventilation/Pumping                    Null     9
#> 28   ET9.M2              Ventilation/Pumping                     Low     9
#> 29   ET9.M3              Ventilation/Pumping                    High     9
#> 30  ET10.M1                     Burrow width                    None    10
#> 31  ET10.M2                     Burrow width                  Narrow    10
#> 32  ET10.M3                     Burrow width            Intermediate    10
#> 33  ET10.M4                     Burrow width                    Wide    10
#> 34  ET11.M1           Endo-3D structure type                    None    11
#> 35  ET11.M2           Endo-3D structure type          Chimney/Funnel    11
#> 36  ET11.M3           Endo-3D structure type                    Tube    11
#> 37  ET11.M4           Endo-3D structure type        IJ-shaped burrow    11
#> 38  ET11.M5           Endo-3D structure type        UY-shaped burrow    11
#> 39  ET11.M6           Endo-3D structure type         Branched burrow    11
#> 40  ET11.M7           Endo-3D structure type      Anastomosed burrow    11
#> 41  ET12.M1          Endo-3D structure depth                    None    12
#> 42  ET12.M2          Endo-3D structure depth                     0-5    12
#> 43  ET12.M3          Endo-3D structure depth                    5-15    12
#> 44  ET12.M4          Endo-3D structure depth                   15-30    12
#> 45  ET12.M5          Endo-3D structure depth                     >30    12
#> 46  ET13.M1            Epi-3D structure type                    None    13
#> 47  ET13.M2            Epi-3D structure type                     Mat    13
#> 48  ET13.M3            Epi-3D structure type                   Mound    13
#> 49  ET13.M4            Epi-3D structure type Tube/Tubular protrusion    13
#> 50  ET13.M5            Epi-3D structure type                   Shell    13
#> 51  ET13.M6            Epi-3D structure type           Stalk/Feather    13
#> 52  ET13.M7            Epi-3D structure type            Protuberance    13
#> 53  ET14.M1       Epi-3D structure extension                    None    14
#> 54  ET14.M2       Epi-3D structure extension        Isolated/Clumped    14
#> 55  ET14.M3       Epi-3D structure extension                Mat/Lawn    14
#> 56  ET14.M4       Epi-3D structure extension             Simple reef    14
#> 57  ET14.M5       Epi-3D structure extension            Complex reef    14
#> 58  ET15.M1            Epi-3D structure size                    None    15
#> 59  ET15.M2            Epi-3D structure size                      <1    15
#> 60  ET15.M3            Epi-3D structure size                     1-3    15
#> 61  ET15.M4            Epi-3D structure size                    3-10    15
#> 62  ET15.M5            Epi-3D structure size                   10-20    15
#> 63  ET15.M6            Epi-3D structure size                   20-50    15
#> 64  ET15.M7            Epi-3D structure size                     >50    15
#> 65   RT1.M1                        Body mass                  <0.001     1
#> 66   RT1.M2                        Body mass             0.001-0.010     1
#> 67   RT1.M3                        Body mass             0.010-0.100     1
#> 68   RT1.M4                        Body mass             0.100-1.000     1
#> 69   RT1.M5                        Body mass                  >1.000     1
#> 70   RT2.M1                      Body length                      <1     2
#> 71   RT2.M2                      Body length                     1-3     2
#> 72   RT2.M3                      Body length                    3-10     2
#> 73   RT2.M4                      Body length                   10-20     2
#> 74   RT2.M5                      Body length                   20-50     2
#> 75   RT3.M1                  Body resistance                Very low     3
#> 76   RT3.M2                  Body resistance                     Low     3
#> 77   RT3.M3                  Body resistance            Intermediate     3
#> 78   RT3.M4                  Body resistance                    High     3
#> 79   RT3.M5                  Body resistance               Very high     3
#> 80   RT4.M1                         Motility                 Sessile     4
#> 81   RT4.M2                         Motility              Tubicolous     4
#> 82   RT4.M3                         Motility                 Crawler     4
#> 83   RT4.M4                         Motility                 Swimmer     4
#> 84   RT5.M1       Burrowing/Sheltering depth                       0     5
#> 85   RT5.M2       Burrowing/Sheltering depth                     0-5     5
#> 86   RT5.M3       Burrowing/Sheltering depth                    5-15     5
#> 87   RT5.M4       Burrowing/Sheltering depth                     >15     5
#> 88   RT6.M1                     Feeding type          Deposit feeder     6
#> 89   RT6.M2                     Feeding type       Suspension feeder     6
#> 90   RT6.M3                     Feeding type        Herbivore/Grazer     6
#> 91   RT6.M4                     Feeding type     Carnivore/Scavenger     6
#> 92   RT7.M1                        Life span                      <1     7
#> 93   RT7.M2                        Life span                     1-3     7
#> 94   RT7.M3                        Life span                    3-10     7
#> 95   RT7.M4                        Life span                   10-20     7
#> 96   RT7.M5                        Life span                     >20     7
#> 97   RT8.M1                  Age at maturity                      <1     8
#> 98   RT8.M2                  Age at maturity                     1-3     8
#> 99   RT8.M3                  Age at maturity                     3-5     8
#> 100  RT9.M1           Reproductive frequency         Sexual seasonal     9
#> 101  RT9.M2           Reproductive frequency       Sexual continuous     9
#> 102 RT10.M1                    Fertilisation            Broadcasting    10
#> 103 RT10.M2                    Fertilisation            Spermcasting    10
#> 104 RT10.M3                    Fertilisation                 Pairing    10
#> 105 RT11.M1                 Annual fecundity                   <10e2    11
#> 106 RT11.M2                 Annual fecundity               10e2-10e3    11
#> 107 RT11.M3                 Annual fecundity               10e3-10e4    11
#> 108 RT11.M4                 Annual fecundity               10e4-10e5    11
#> 109 RT11.M5                 Annual fecundity               10e5-10e6    11
#> 110 RT11.M6                 Annual fecundity                   >10e6    11
#> 111 RT12.M1                   Offspring type                     Egg    12
#> 112 RT12.M2                   Offspring type                   Larva    12
#> 113 RT12.M3                   Offspring type                Juvenile    12
#> 114 RT13.M1                   Offspring size                    <0.1    13
#> 115 RT13.M2                   Offspring size                 0.1-0.5    13
#> 116 RT13.M3                   Offspring size                 0.5-1.5    13
#> 117 RT13.M4                   Offspring size                   1.5-5    13
#> 118 RT14.M1             Offspring protection                    None    14
#> 119 RT14.M2             Offspring protection                     Gel    14
#> 120 RT14.M3             Offspring protection                 Capsule    14
#> 121 RT14.M4             Offspring protection                Burrying    14
#> 122 RT14.M5             Offspring protection        Bearing/Brooding    14
#> 123 RT15.M1            Offspring development          Planktotrophic    15
#> 124 RT15.M2            Offspring development          Lecithotrophic    15
#> 125 RT15.M3            Offspring development    Mixed planktotrophic    15
#> 126 RT15.M4            Offspring development    Mixed lecithotrophic    15
#> 127 RT15.M5            Offspring development                Internal    15
#> 128 RT16.M1 Offspring benthic stage duration                    Null    16
#> 129 RT16.M2 Offspring benthic stage duration                     <15    16
#> 130 RT16.M3 Offspring benthic stage duration                   15-30    16
#> 131 RT16.M4 Offspring benthic stage duration                   30-60    16
#> 132 RT16.M5 Offspring benthic stage duration                     >60    16
#> 133 RT17.M1 Offspring pelagic stage duration                    Null    17
#> 134 RT17.M2 Offspring pelagic stage duration                     <15    17
#> 135 RT17.M3 Offspring pelagic stage duration                   15-30    17
#> 136 RT17.M4 Offspring pelagic stage duration                   30-60    17
#> 137 RT17.M5 Offspring pelagic stage duration                     >60    17
#>            value     score units
#> 1   0.000000e+00 1.0000000    cm
#> 2   2.500000e+00 0.7500000    cm
#> 3   1.000000e+01 0.5000000    cm
#> 4   2.250000e+01 0.2500000    cm
#> 5   3.000000e+01 0.0000000    cm
#> 6   0.000000e+00 0.0000000     -
#> 7   5.000000e-01 0.5000000     -
#> 8   1.000000e+00 1.0000000     -
#> 9   0.000000e+00 0.0000000     -
#> 10  5.000000e-01 0.5000000     -
#> 11  1.000000e+00 1.0000000     -
#> 12  0.000000e+00 0.0000000     -
#> 13  5.000000e-01 0.5000000     -
#> 14  1.000000e+00 1.0000000     -
#> 15  0.000000e+00 0.0000000     -
#> 16  5.000000e-01 0.5000000     -
#> 17  1.000000e+00 1.0000000     -
#> 18  0.000000e+00 0.0000000     -
#> 19  5.000000e-01 0.5000000     -
#> 20  1.000000e+00 1.0000000     -
#> 21  0.000000e+00 0.0000000     -
#> 22  5.000000e-01 0.5000000     -
#> 23  1.000000e+00 1.0000000     -
#> 24  0.000000e+00 0.0000000     -
#> 25  5.000000e-01 0.5000000     -
#> 26  1.000000e+00 1.0000000     -
#> 27  0.000000e+00 0.0000000     -
#> 28  5.000000e-01 0.5000000     -
#> 29  1.000000e+00 1.0000000     -
#> 30  0.000000e+00 0.0000000    mm
#> 31  5.000000e+00 0.3333333    mm
#> 32  7.500000e+00 0.6666667    mm
#> 33  1.250000e+01 1.0000000    mm
#> 34  0.000000e+00 0.0000000     -
#> 35  1.666667e-01 0.1666667     -
#> 36  3.333333e-01 0.3333333     -
#> 37  5.000000e-01 0.5000000     -
#> 38  6.666667e-01 0.6666667     -
#> 39  8.333333e-01 0.8333333     -
#> 40  1.000000e+00 1.0000000     -
#> 41  0.000000e+00 0.0000000    cm
#> 42  2.500000e+00 0.2500000    cm
#> 43  1.000000e+01 0.5000000    cm
#> 44  2.250000e+01 0.7500000    cm
#> 45  3.000000e+01 1.0000000    cm
#> 46  0.000000e+00 0.0000000     -
#> 47  1.666667e-01 0.1666667     -
#> 48  3.333333e-01 0.3333333     -
#> 49  5.000000e-01 0.5000000     -
#> 50  6.666667e-01 0.6666667     -
#> 51  8.333333e-01 0.8333333     -
#> 52  1.000000e+00 1.0000000     -
#> 53  0.000000e+00 0.0000000     -
#> 54  2.500000e-01 0.2500000     -
#> 55  5.000000e-01 0.5000000     -
#> 56  7.500000e-01 0.7500000     -
#> 57  1.000000e+00 1.0000000     -
#> 58  0.000000e+00 0.0000000    mm
#> 59  1.000000e+00 0.1666667    mm
#> 60  2.000000e+00 0.3333333    mm
#> 61  6.500000e+00 0.5000000    mm
#> 62  1.500000e+01 0.6666667    mm
#> 63  3.500000e+01 0.8333333    mm
#> 64  5.000000e+01 1.0000000    mm
#> 65  5.000000e-04 0.0000000 gADWt
#> 66  5.000000e-03 0.2500000 gADWt
#> 67  5.000000e-02 0.5000000 gADWt
#> 68  5.000000e-01 0.7500000 gADWt
#> 69  5.000000e+00 1.0000000 gADWt
#> 70  5.000000e-01 0.0000000    mm
#> 71  2.000000e+00 0.2500000    mm
#> 72  6.500000e+00 0.5000000    mm
#> 73  1.500000e+01 0.7500000    mm
#> 74  3.500000e+01 1.0000000    mm
#> 75  0.000000e+00 1.0000000     -
#> 76  2.500000e-01 0.7500000     -
#> 77  5.000000e-01 0.5000000     -
#> 78  7.500000e-01 0.2500000     -
#> 79  1.000000e+00 0.0000000     -
#> 80  0.000000e+00 1.0000000     -
#> 81  3.333333e-01 0.6666667     -
#> 82  6.666667e-01 0.3333333     -
#> 83  1.000000e+00 0.0000000     -
#> 84  0.000000e+00 1.0000000    cm
#> 85  2.500000e+00 0.7500000    cm
#> 86  1.000000e+01 0.5000000    cm
#> 87  2.000000e+01 0.2500000    cm
#> 88  0.000000e+00 0.0000000     -
#> 89  3.333333e-01 0.3333333     -
#> 90  6.666667e-01 0.6666667     -
#> 91  1.000000e+00 1.0000000     -
#> 92  5.000000e-01 0.0000000  year
#> 93  2.000000e+00 0.2500000  year
#> 94  6.500000e+00 0.5000000  year
#> 95  1.500000e+01 0.7500000  year
#> 96  3.000000e+01 1.0000000  year
#> 97  5.000000e-01 0.0000000  year
#> 98  2.000000e+00 0.5000000  year
#> 99  4.000000e+00 1.0000000  year
#> 100 0.000000e+00 0.0000000     -
#> 101 1.000000e+00 1.0000000     -
#> 102 0.000000e+00 0.0000000     -
#> 103 5.000000e-01 0.5000000     -
#> 104 1.000000e+00 1.0000000     -
#> 105 5.000000e+01 0.0000000     -
#> 106 4.500000e+02 0.2000000     -
#> 107 4.500000e+03 0.4000000     -
#> 108 4.500000e+04 0.6000000     -
#> 109 4.500000e+05 0.8000000     -
#> 110 1.000000e+06 1.0000000     -
#> 111 0.000000e+00 0.0000000     -
#> 112 5.000000e-01 0.5000000     -
#> 113 1.000000e+00 1.0000000     -
#> 114 5.000000e-02 0.0000000    mm
#> 115 3.000000e-01 0.3333333    mm
#> 116 1.000000e+00 0.6666667    mm
#> 117 3.250000e+00 1.0000000    mm
#> 118 0.000000e+00 0.0000000     -
#> 119 2.500000e-01 0.2500000     -
#> 120 5.000000e-01 0.5000000     -
#> 121 7.500000e-01 0.7500000     -
#> 122 1.000000e+00 1.0000000     -
#> 123 0.000000e+00 0.0000000     -
#> 124 2.500000e-01 0.2500000     -
#> 125 5.000000e-01 0.5000000     -
#> 126 7.500000e-01 0.7500000     -
#> 127 1.000000e+00 1.0000000     -
#> 128 0.000000e+00 0.0000000  days
#> 129 7.500000e+00 0.2500000  days
#> 130 2.250000e+01 0.5000000  days
#> 131 4.500000e+01 0.7500000  days
#> 132 6.000000e+01 1.0000000  days
#> 133 0.000000e+00 0.0000000  days
#> 134 7.500000e+00 0.2500000  days
#> 135 2.250000e+01 0.5000000  days
#> 136 4.500000e+01 0.7500000  days
#> 137 6.000000e+01 1.0000000  days

## ====================================================
## Traits for certain taxa
## ====================================================

# Taxonomy of amphiura and related species in the databases.
getTaxonomy("Amphiura filiformis")
#> [[1]]
#>                   taxon    genus      family         order       class
#> 112 Amphiura filiformis Amphiura Amphiuridae Amphilepidida Ophiuroidea
#>            phylum AphiaID
#> 112 Echinodermata  125080
#> 
getTaxonomy("Amphiura")
#> [[1]]
#>                                 taxon    genus      family         order
#> 109                          Amphiura Amphiura Amphiuridae Amphilepidida
#> 110 Amphiura (Ophiopeltis) securigera Amphiura Amphiuridae Amphilepidida
#> 111                  Amphiura chiajei Amphiura Amphiuridae Amphilepidida
#> 112               Amphiura filiformis Amphiura Amphiuridae Amphilepidida
#>           class        phylum AphiaID
#> 109 Ophiuroidea Echinodermata  123613
#> 110 Ophiuroidea Echinodermata  125195
#> 111 Ophiuroidea Echinodermata  125073
#> 112 Ophiuroidea Echinodermata  125080
#> 

getTaxonInfo("Amphiura filiformis")
#>                  taxon                            trait          modality units
#> 1  Amphiura filiformis    Substratum depth distribution               0-5    cm
#> 2  Amphiura filiformis    Substratum depth distribution              5-15    cm
#> 3  Amphiura filiformis                     Biodiffusion               Low     -
#> 7  Amphiura filiformis                    Biodeposition               Low     -
#> 10 Amphiura filiformis              Ventilation/Pumping              High     -
#> 11 Amphiura filiformis                     Burrow width            Narrow    mm
#> 12 Amphiura filiformis                     Burrow width      Intermediate    mm
#> 13 Amphiura filiformis           Endo-3D structure type    Chimney/Funnel     -
#> 14 Amphiura filiformis           Endo-3D structure type  UY-shaped burrow     -
#> 15 Amphiura filiformis          Endo-3D structure depth               0-5    cm
#> 16 Amphiura filiformis          Endo-3D structure depth              5-15    cm
#> 17 Amphiura filiformis            Epi-3D structure type             Mound     -
#> 18 Amphiura filiformis       Epi-3D structure extension  Isolated/Clumped     -
#> 19 Amphiura filiformis            Epi-3D structure size                <1    mm
#> 20 Amphiura filiformis            Epi-3D structure size               1-3    mm
#> 21 Amphiura filiformis                        Body mass       0.001-0.010 gADWt
#> 22 Amphiura filiformis                      Body length             20-50    mm
#> 23 Amphiura filiformis                  Body resistance               Low     -
#> 24 Amphiura filiformis                         Motility           Crawler     -
#> 25 Amphiura filiformis       Burrowing/Sheltering depth              5-15    cm
#> 26 Amphiura filiformis                     Feeding type    Deposit feeder     -
#> 27 Amphiura filiformis                     Feeding type Suspension feeder     -
#> 28 Amphiura filiformis                        Life span             10-20  year
#> 29 Amphiura filiformis                  Age at maturity               3-5  year
#> 30 Amphiura filiformis           Reproductive frequency   Sexual seasonal     -
#> 31 Amphiura filiformis                    Fertilisation           Pairing     -
#> 32 Amphiura filiformis                 Annual fecundity         10e4-10e5     -
#> 33 Amphiura filiformis                   Offspring type               Egg     -
#> 34 Amphiura filiformis                   Offspring size              <0.1    mm
#> 36 Amphiura filiformis            Offspring development    Planktotrophic     -
#> 38 Amphiura filiformis Offspring pelagic stage duration             30-60  days
#>    proportion
#> 1        0.75
#> 2        0.25
#> 3        1.00
#> 7        1.00
#> 10       1.00
#> 11       0.50
#> 12       0.50
#> 13       0.50
#> 14       0.50
#> 15       0.50
#> 16       0.50
#> 17       1.00
#> 18       1.00
#> 19       0.50
#> 20       0.50
#> 21       1.00
#> 22       1.00
#> 23       1.00
#> 24       1.00
#> 25       1.00
#> 26       0.50
#> 27       0.50
#> 28       1.00
#> 29       1.00
#> 30       1.00
#> 31       1.00
#> 32       1.00
#> 33       1.00
#> 34       1.00
#> 36       1.00
#> 38       1.00
getTaxonInfo("Amphiura filiformis", trait=Traits_cefas)
#>                  taxon                      trait                  modality
#> 1  Amphiura filiformis               Maximum size                   101-200
#> 2  Amphiura filiformis                 Morphology               Exoskeleton
#> 3  Amphiura filiformis                   Lifespan                       >10
#> 4  Amphiura filiformis   Egg development location            Sexual_pelagic
#> 5  Amphiura filiformis Larva development location    Pelagic_planktotrophic
#> 6  Amphiura filiformis               Living habit               Free_living
#> 7  Amphiura filiformis               Living habit Crevice_hole_under_stones
#> 8  Amphiura filiformis          Sediment position  Shallow_infauna_0_to_5cm
#> 9  Amphiura filiformis               Feeding mode                Suspension
#> 10 Amphiura filiformis               Feeding mode           Surface_deposit
#> 11 Amphiura filiformis                   Mobility         Crawl_creep_climb
#> 12 Amphiura filiformis                   Mobility                  Burrower
#> 13 Amphiura filiformis          Bioturbation mode          Diffusive_mixing
#> 14 Amphiura filiformis          Bioturbation mode        Surface_deposition
#>    units proportion
#> 1     mm       1.00
#> 2      -       1.00
#> 3  years       1.00
#> 4      -       1.00
#> 5      -       1.00
#> 6      -       0.75
#> 7      -       0.25
#> 8     cm       1.00
#> 9      -       0.50
#> 10     -       0.50
#> 11     -       0.50
#> 12     -       0.50
#> 13     -       0.50
#> 14     -       0.50

# there is no direct information on the traits of this species in Traits_cefas
getTaxonInfo("Amphiura filiformis", trait=Traits_cefas, NULL)
#> [1] taxon      trait      modality   units      proportion
#> <0 rows> (or 0-length row.names)

## ====================================================
## Northsea dataset
## ====================================================

# show its contents
metadata(MWTL)          
#> $title
#> [1] "The MWTL Northsea macrobenthos data (1995 - 2018)"
#> 
#> $originator
#> [1] "Ministerie van Infrastructuur en Milieu, Rijkswaterstaat Centrale Informatievoorziening (RWS, CIV)"
#> 
#> $implemented
#> [1] "Karline Soetaert, in the frame of the EMODnet biology project"
#> 
metadata(MWTL$density)
#>             name                                description          units
#> 1        station                               station name               
#> 2           date                    sampling date, a string               
#> 3          taxon taxon name, checked by worms, and adjusted               
#> 4        density                      species total density individuals/m2
#> 5        biomass          species total ash-free dry weight       gAFDW/m2
#> 6 taxon.original                        original taxon name